Support ingress gateways
This creates all of the necessary templates to enable ingress gateways
in the Helm chart. By making use of `ingressGateways.defaults` and
`ingressGateways.gateways` setting, it's possible to enable multiple
ingress gateways. Names must be provided for each and they must be
unique.

All fields defined in the `ingressGateways.defaults` may be overridden
by defining that field on the relevant `ingressGateways.gateways`
entry, with the exception of annotations. Annotations will apply
both the default annotations and any gateway specific annotations.

In an effort to reduce the intrusiveness of the installation, this
feature uses a role and a rolebinding rather than clusterrole/binding.

TODO: Testing.